requirements.txt
numpy
pandas
python-dotenv
ultralytics
opencv-python
google-genai